--- a/src/NuGet.Core/NuGet.Packaging/RuntimeModel/IObjectWriter.cs
+++ b/src/NuGet.Core/NuGet.Packaging/RuntimeModel/IObjectWriter.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Generic;
 
 namespace NuGet.RuntimeModel
@@ -12,6 +13,17 @@ namespace NuGet.RuntimeModel
     public interface IObjectWriter
     {
         /// <summary>
+        /// Writes the start of the root object or an object in an array.
+        ///
+        /// This new object becomes the scope for all other method calls until either WriteObjectStart
+        /// is called again to start a new nested object or WriteObjectEnd is called.
+        ///
+        /// Every call to WriteObjectStart must be balanced by a corresponding call to WriteObjectEnd.
+        /// </summary>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
+        void WriteObjectStart();
+
+        /// <summary>
         /// Writes the start of a nested object.
         ///
         /// This new object becomes the scope for all other method calls until either WriteObjectStart
@@ -19,7 +31,9 @@ namespace NuGet.RuntimeModel
         ///
         /// Every call to WriteObjectStart must be balanced by a corresponding call to WriteObjectEnd.
         /// </summary>
-        /// <param name="name">The name of the object.  Throws if <c>null</c>.</param>
+        /// <param name="name">The name of the object.</param>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="name" /> is <c>null</c>.</exception>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
         void WriteObjectStart(string name);
 
         /// <summary>
@@ -30,42 +44,53 @@ namespace NuGet.RuntimeModel
         ///
         /// Every call to WriteObjectStart must be balanced by a corresponding call to WriteObjectEnd.
         /// </summary>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
         void WriteObjectEnd();
 
         /// <summary>
         /// Writes a name-value pair.
         /// </summary>
-        /// <param name="name">The name of the datum.  Throws if <c>null</c>.</param>
+        /// <param name="name">The name of the datum.</param>
         /// <param name="value">The datum.</param>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="name" /> is <c>null</c>.</exception>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
         void WriteNameValue(string name, int value);
 
         /// <summary>
         /// Writes a name-value pair.
         /// </summary>
-        /// <param name="name">The name of the datum.  Throws if <c>null</c>.</param>
+        /// <param name="name">The name of the datum.</param>
         /// <param name="value">The datum.</param>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="name" /> is <c>null</c>.</exception>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
         void WriteNameValue(string name, bool value);
 
         /// <summary>
         /// Writes a name-value pair.
         /// </summary>
-        /// <param name="name">The name of the datum.  Throws if <c>null</c>.</param>
+        /// <param name="name">The name of the datum.</param>
         /// <param name="value">The datum.</param>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="name" /> is <c>null</c>.</exception>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
         void WriteNameValue(string name, string value);
 
         /// <summary>
         /// Writes a name-collection pair.
         /// </summary>
-        /// <param name="name">The name of the data.  Throws if <c>null</c>.</param>
+        /// <param name="name">The name of the data.</param>
         /// <param name="values">The data.</param>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="name" /> is <c>null</c>.</exception>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
         void WriteNameArray(string name, IEnumerable<string> values);
 
         /// <summary>
         /// Writes the start of an array.
-        /// The new object becomes the scope of all other methods until WriteObjectInArrayStart is called to start a new object in the array, or WriteArrayEnd is called.
+        /// The new object becomes the scope of all other methods until WriteArrayStart is called to start a new object in the array, or WriteArrayEnd is called.
         /// Every call to WriteArrayStart needs to be balanced with a corresponding call to WriteArrayEnd and not WriteObjectEnd.
         /// </summary>
         /// <param name="name">The array name</param>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="name" /> is <c>null</c>.</exception>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
         void WriteArrayStart(string name);
 
         /// <summary>
@@ -76,18 +101,7 @@ namespace NuGet.RuntimeModel
         ///
         /// Every call to WriteArrayStart needs to be balanced with a corresponding call to WriteArrayEnd and not WriteObjectEnd.
         /// </summary>
+        /// <exception cref="ObjectDisposedException">Thrown if this object is disposed.</exception>
         void WriteArrayEnd();
-
-        /// <summary>
-        /// Writes the start of a nested object in array.
-        ///
-        /// This new object becomes the scope for all other method calls until either WriteObjectInArrayStart
-        /// is called again to start a new nested object or WriteObjectEnd is called.
-        ///
-        /// Every call to WriteObjectInArrayStart must be balanced by a corresponding call to WriteObjectEnd.
-        /// </summary>
-        void WriteObjectInArrayStart();
-
-
     }
 }