… src/NuGet.Core/NuGet.ProjectModel/LockFile/LockFileFormat.cs
… test/NuGet.Core.Tests/NuGet.Packaging.Test/RuntimeModelTests/JsonObjectWriterTests.cs
… test/NuGet.Core.Tests/NuGet.ProjectModel.Test/DependencyGraphSpecTests.cs
… test/NuGet.Core.Tests/NuGet.ProjectModel.Test/HashObjectWriterTests.cs
… test/NuGet.Core.Tests/NuGet.ProjectModel.Test/PackageSpecCloningTests.cs
… test/NuGet.Core.Tests/NuGet.ProjectModel.Test/PackageSpecTestUtility.cs
… test/NuGet.Core.Tests/NuGet.ProjectModel.Test/PackageSpecWriterTests.cs