## Bug

Fixes: https://github.com/NuGet/Home/issues/9031
Regression: No

## Fix

Details: `DependencyGraphSpec.Save(…)` saves a dependency graph spec (DG spec) to disk.  The previous implementation serialized the `JObject` representation of the DG spec into a string and then wrote the string to disk.  This intermediate step of serializing to a string is both unnecessary and a performance issue.  As project and solution DG specs grow beyond 80 KB, these strings will be allocated on the large object heap (LOH).  This has terrible [performance implications](https://docs.microsoft.com/en-us/dotnet/standard/garbage-collection/large-object-heap#loh-performance-implications) given that NuGet restore is triggered automatically in many scenarios (i.e.:  build, rebuild, debug, code analysis, etc.).

## Testing/Validation

Tests Added: Yes
Validation:  Using the NuGet.Client solution as a case study, I profiled performance 5 times before and 5 times after this change.  More iterations than 5 would have been great, but it's time consuming.

Deleting only *.nuget.dgspec.json files before each test (to force calls to `DependencyGraphSpec.Save(…)`), I observed the following "no-op-ish" restore performance changes in Visual Studio.  I say "no-op-ish" because there's no real scenario where only these files are deleted; however, the scenario is no-op in all other respects:  a full restore was performed previous to testing.  It's also important to note that these numbers are based on _all) activity in VS during a restore operation not just NuGet activity; there are many variables which I cannot control within VS.

All numbers below are averages.

| Metric |  Before | After | Delta | % |
| ---  | --- | --- | --- | --- |
| "no-op-ish" restore time |  2,069.96 ms | 1,925.08 ms | -144.88 ms | -7 % |
| heap allocations | 331.61 MB | 308.55 MB | -23.06 MB | -6.95% |
| LOH allocations | 3.80 MB | 0 MB | -3.80 MB | -100% |
| total GC pause | 1,324.60 ms | 1,172.94 ms | -151.66 ms | -11.45% |

A real-life no-op restore --- that is one without me deleting *.nuget.dgspec.json files outside of Visual Studio --- allocated 776,496 bytes on the LOH.  With this change it's 0.